+++ plugins/Report/src/Controller/ReportsController.php
@@ -45,12 +45,22 @@ class ReportsController extends AppController
         $this->set('contentHeader', $header);
     }
 
+    public function getInstitutionStatusOptions($module)
+    {
+        $options = [
+                'Active' => __('Active'),
+                'Inactive' => __('Inactive')
+            ];
+            return $options;
+    }
+
     public function getFeatureOptions($module)
     {
         $options = [];
         if ($module == 'Directory') {
             $options = [
                 'Report.Directory' => __('User Default Identity'),
+                'Report.Users'     => __('User List')
             ];
         } elseif ($module == 'Institutions') {
             $options = [
@@ -64,6 +74,7 @@ class ReportsController extends AppController
                 'Report.InstitutionStaff' => __('Staff'),
                 'Report.StudentAbsences' => __('Student Absence'),
                 'Report.StudentAttendanceSummary' => __('Student Attendance Summary'),
+                'Report.InstitutionSummaryReport' => __('Institution Summary Report'),
                 'Report.BodyMasses' => __('Student Body Masses'),
                 // 'Report.StaffAbsences' => __('Staff Absence'),
                 'Report.StaffAttendances' => __('Staff Attendance'),
@@ -79,7 +90,9 @@ class ReportsController extends AppController
                 'Report.SpecialNeedsFacilities' => __('Special Needs Facilities'),
                 'Report.InstitutionCommittees' => __('Committees'),
                 'Report.InstitutionSubjectsClasses' => __('Subjects/Classes'),
-                'Report.ClassAttendanceMarkedSummaryReport' => __('Class Attendance Marked Summary Report')
+                'Report.ClassAttendanceMarkedSummaryReport' => __('Class Attendance Marked Summary Report'),
+                'Report.Income' => __('Income Report'),
+                'Report.Expenditure' => __('Expenditure Report')
 				
             ];
         } elseif ($module == 'Students') {
@@ -112,7 +125,8 @@ class ReportsController extends AppController
                 'Report.StaffSystemUsage' => __('System Usage'),
                 'Report.StaffTrainingReports' => __('Training Courses Report'),
                 'Report.StaffLeaveReport' => __('Staff Leave'),
-                'Report.StaffPositions' => __('Position Summary Report')
+                'Report.StaffPositions' => __('Position Summary Report'),
+                'Report.StaffDuties' => __('Duties Report')
 				
             ];
         } elseif ($module == 'Textbooks') {
@@ -197,6 +211,7 @@ class ReportsController extends AppController
                 'ReportProgress.total_records'
             );
             $ReportProgress = TableRegistry::get('Report.ReportProgress');
+            echo '<pre>';print_r($ReportProgress);die;
             if (!empty($ids)) {
                 $results = $ReportProgress
                     ->find()